backend/requirements.txt
alembic==1.3.1
appdirs==1.4.3
attrs==19.3.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.2
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
prometheus-client==0.7.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
redis==3.3.11
requests==2.22.0
requests-toolbelt==0.9.1
rq==1.1.0
six==1.13.0
SQLAlchemy==1.3.11
SQLAlchemy-Utils==0.35.0
urllib3==1.25.7
Werkzeug==0.16.0
zeep==3.4.0
